+++ lib/src/amazon/voices/voice_model.g.dart
@@ -7,6 +7,8 @@ part of 'voice_model.dart';
 // **************************************************************************
 
 VoiceAmazon _$VoiceAmazonFromJson(Map<String, dynamic> json) => VoiceAmazon(
+      provider: json['provider'] as String? ?? TtsProviders.amazon,
+      engines: VoiceAmazon._toEngines(json['SupportedEngines'] as List),
       code: json['Id'] as String,
       voiceType: VoiceAmazon._toVoiceType(json['SupportedEngines'] as List),
       name: json['Id'] as String,