+++ lib/src/cloud_text_to_speech_base.dart
@@ -1,3 +1,4 @@
+export '/src/common/tts/tts_providers.dart';
 export '/src/common/locale/locale_model.dart';
 export '/src/google/common/init.dart';
 export '/src/google/audio/audio.dart';