+++ lib/src/universal/voices/voice_model.dart
@@ -1,7 +1,10 @@
 import 'package:cloud_text_to_speech/src/common/locale/locale_model.dart';
 
 class VoiceUniversal {
+  String provider;
+  List<String> engines;
   String code;
+  @Deprecated("Use engines instead")
   String voiceType;
   String name;
   String nativeName;
@@ -9,6 +12,8 @@ class VoiceUniversal {
   VoiceLocale locale;
 
   VoiceUniversal({
+    required this.provider,
+    required this.engines,
     required this.code,
     required this.voiceType,
     required this.name,